+++ tests/integration/test_models/standard_output_tests.py
@@ -631,7 +631,7 @@ class CurrentTests(BaseOutputTest):
                 axis=0,
             ),
             self.i_cell / self.l_n,
-            decimal=3,
+            decimal=4,
         )
         np.testing.assert_array_almost_equal(
             np.mean(
@@ -640,7 +640,7 @@ class CurrentTests(BaseOutputTest):
                 axis=0,
             ),
             -self.i_cell / self.l_p,
-            decimal=3,
+            decimal=4,
         )
 
     def test_conservation(self):