+++ tests/integration/test_models/test_full_battery_models/test_lithium_ion/test_newman_tobias.py
@@ -1,5 +1,5 @@
 #
-# Tests for the lithium-ion DFN model
+# Tests for the lithium-ion Newman-Tobias model
 #
 import pybamm
 import tests
@@ -8,22 +8,18 @@ import numpy as np
 import unittest
 
 
-class TestDFN(unittest.TestCase):
+class TestNewmanTobias(unittest.TestCase):
     def test_basic_processing(self):
         options = {"thermal": "isothermal"}
-        model = pybamm.lithium_ion.DFN(options)
-        # use Ecker parameters for nonlinear diffusion
-        param = pybamm.ParameterValues(chemistry=pybamm.parameter_sets.Ecker2015)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         var = pybamm.standard_spatial_vars
         var_pts = {var.x_n: 10, var.x_s: 10, var.x_p: 10, var.r_n: 5, var.r_p: 5}
-        modeltest = tests.StandardModelTest(
-            model, parameter_values=param, var_pts=var_pts
-        )
+        modeltest = tests.StandardModelTest(model, var_pts=var_pts)
         modeltest.test_all()
 
     def test_basic_processing_1plus1D(self):
         options = {"current collector": "potential pair", "dimensionality": 1}
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         var = pybamm.standard_spatial_vars
         var_pts = {
             var.x_n: 5,
@@ -39,7 +35,7 @@ class TestDFN(unittest.TestCase):
 
     def test_basic_processing_2plus1D(self):
         options = {"current collector": "potential pair", "dimensionality": 2}
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         var = pybamm.standard_spatial_vars
         var_pts = {
             var.x_n: 5,
@@ -55,7 +51,7 @@ class TestDFN(unittest.TestCase):
 
     def test_optimisations(self):
         options = {"thermal": "isothermal"}
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         optimtest = tests.OptimisationsTest(model)
 
         original = optimtest.evaluate_model()
@@ -65,14 +61,32 @@ class TestDFN(unittest.TestCase):
         np.testing.assert_array_almost_equal(original, to_python)
 
     def test_set_up(self):
-        model = pybamm.lithium_ion.DFN()
+        model = pybamm.lithium_ion.NewmanTobias()
         optimtest = tests.OptimisationsTest(model)
         optimtest.set_up_model(to_python=True)
         optimtest.set_up_model(to_python=False)
 
+    def test_particle_fickian(self):
+        options = {"particle": "Fickian diffusion"}
+        model = pybamm.lithium_ion.NewmanTobias(options)
+        modeltest = tests.StandardModelTest(model)
+        modeltest.test_all()
+
+    def test_particle_quadratic(self):
+        options = {"particle": "quadratic profile"}
+        model = pybamm.lithium_ion.NewmanTobias(options)
+        modeltest = tests.StandardModelTest(model)
+        modeltest.test_all()
+
+    def test_particle_quartic(self):
+        options = {"particle": "quartic profile"}
+        model = pybamm.lithium_ion.NewmanTobias(options)
+        modeltest = tests.StandardModelTest(model)
+        modeltest.test_all()
+
     def test_full_thermal(self):
         options = {"thermal": "x-full"}
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         var = pybamm.standard_spatial_vars
         var_pts = {var.x_n: 10, var.x_s: 10, var.x_p: 10, var.r_n: 5, var.r_p: 5}
         modeltest = tests.StandardModelTest(model, var_pts=var_pts)
@@ -80,33 +94,15 @@ class TestDFN(unittest.TestCase):
 
     def test_lumped_thermal(self):
         options = {"thermal": "lumped"}
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         var = pybamm.standard_spatial_vars
         var_pts = {var.x_n: 10, var.x_s: 10, var.x_p: 10, var.r_n: 5, var.r_p: 5}
         modeltest = tests.StandardModelTest(model, var_pts=var_pts)
         modeltest.test_all()
 
-    def test_particle_uniform(self):
-        options = {"particle": "uniform profile"}
-        model = pybamm.lithium_ion.DFN(options)
-        modeltest = tests.StandardModelTest(model)
-        modeltest.test_all()
-
-    def test_particle_quadratic(self):
-        options = {"particle": "quadratic profile"}
-        model = pybamm.lithium_ion.DFN(options)
-        modeltest = tests.StandardModelTest(model)
-        modeltest.test_all()
-
-    def test_particle_quartic(self):
-        options = {"particle": "quartic profile"}
-        model = pybamm.lithium_ion.DFN(options)
-        modeltest = tests.StandardModelTest(model)
-        modeltest.test_all()
-
     def test_loss_active_material(self):
         options = {"particle cracking": "none", "loss of active material": "none"}
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         chemistry = pybamm.parameter_sets.Ai2020
         parameter_values = pybamm.ParameterValues(chemistry=chemistry)
         modeltest = tests.StandardModelTest(model, parameter_values=parameter_values)
@@ -117,7 +113,7 @@ class TestDFN(unittest.TestCase):
             "particle cracking": "no cracking",
             "loss of active material": "negative",
         }
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         chemistry = pybamm.parameter_sets.Ai2020
         parameter_values = pybamm.ParameterValues(chemistry=chemistry)
         modeltest = tests.StandardModelTest(model, parameter_values=parameter_values)
@@ -128,7 +124,7 @@ class TestDFN(unittest.TestCase):
             "particle cracking": "no cracking",
             "loss of active material": "positive",
         }
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         chemistry = pybamm.parameter_sets.Ai2020
         parameter_values = pybamm.ParameterValues(chemistry=chemistry)
         modeltest = tests.StandardModelTest(model, parameter_values=parameter_values)
@@ -139,7 +135,7 @@ class TestDFN(unittest.TestCase):
             "particle cracking": "no cracking",
             "loss of active material": "both",
         }
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         chemistry = pybamm.parameter_sets.Ai2020
         parameter_values = pybamm.ParameterValues(chemistry=chemistry)
         modeltest = tests.StandardModelTest(model, parameter_values=parameter_values)
@@ -147,121 +143,89 @@ class TestDFN(unittest.TestCase):
 
     def test_surface_form_differential(self):
         options = {"surface form": "differential"}
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         modeltest = tests.StandardModelTest(model)
         modeltest.test_all()
 
     def test_surface_form_algebraic(self):
         options = {"surface form": "algebraic"}
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         modeltest = tests.StandardModelTest(model)
         modeltest.test_all()
 
-    def test_particle_distribution_in_x(self):
-        model = pybamm.lithium_ion.DFN()
-        param = model.default_parameter_values
-        L_n = model.param.L_n
-        L_p = model.param.L_p
-        L = model.param.L_x
 
-        def negative_radius(x):
-            return (1 + x / L_n) * 1e-5
-
-        def positive_radius(x):
-            return (1 + (x - L_p) / (L - L_p)) * 1e-5
-
-        param["Negative particle radius [m]"] = negative_radius
-        param["Positive particle radius [m]"] = positive_radius
-        # Only get 3dp of accuracy in some tests at 1C with particle distribution
-        # TODO: investigate if there is a bug or some way to improve the
-        # implementation
-        param["Current function [A]"] = 0.5 * param["Nominal cell capacity [A.h]"]
-        modeltest = tests.StandardModelTest(model, parameter_values=param)
-        modeltest.test_all()
-
-    def test_well_posed_irreversible_plating_with_porosity(self):
-        options = {
-            "lithium plating": "irreversible",
-            "lithium plating porosity change": "true",
-        }
-        model = pybamm.lithium_ion.DFN(options)
-        param = pybamm.ParameterValues(chemistry=pybamm.parameter_sets.Yang2017)
-        modeltest = tests.StandardModelTest(model, parameter_values=param)
-        modeltest.test_all()
-
-
-class TestDFNWithSEI(unittest.TestCase):
+class TestNewmanTobiasWithSEI(unittest.TestCase):
     def test_well_posed_constant(self):
         options = {"SEI": "constant"}
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         modeltest = tests.StandardModelTest(model)
         modeltest.test_all()
 
     def test_well_posed_reaction_limited(self):
         options = {"SEI": "reaction limited"}
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         modeltest = tests.StandardModelTest(model)
         modeltest.test_all()
 
     def test_well_posed_reaction_limited_average_film_resistance(self):
         options = {"SEI": "reaction limited", "SEI film resistance": "average"}
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         modeltest = tests.StandardModelTest(model)
         modeltest.test_all()
 
     def test_well_posed_solvent_diffusion_limited(self):
         options = {"SEI": "solvent-diffusion limited"}
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         modeltest = tests.StandardModelTest(model)
         modeltest.test_all()
 
     def test_well_posed_electron_migration_limited(self):
         options = {"SEI": "electron-migration limited"}
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         modeltest = tests.StandardModelTest(model)
         modeltest.test_all()
 
     def test_well_posed_interstitial_diffusion_limited(self):
         options = {"SEI": "interstitial-diffusion limited"}
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         modeltest = tests.StandardModelTest(model)
         modeltest.test_all()
 
     def test_well_posed_ec_reaction_limited(self):
         options = {"SEI": "ec reaction limited", "SEI porosity change": "true"}
-        model = pybamm.lithium_ion.DFN(options)
+        model = pybamm.lithium_ion.NewmanTobias(options)
         modeltest = tests.StandardModelTest(model)
         modeltest.test_all()
 
 
-class TestDFNWithCrack(unittest.TestCase):
+class TestNewmanTobiasWithCrack(unittest.TestCase):
     def test_well_posed_no_cracking(self):
-        options = {"particle": "Fickian diffusion", "particle cracking": "no cracking"}
-        model = pybamm.lithium_ion.DFN(options)
+        options = {"particle cracking": "no cracking"}
+        model = pybamm.lithium_ion.NewmanTobias(options)
         chemistry = pybamm.parameter_sets.Ai2020
         parameter_values = pybamm.ParameterValues(chemistry=chemistry)
         modeltest = tests.StandardModelTest(model, parameter_values=parameter_values)
         modeltest.test_all()
 
     def test_well_posed_negative_cracking(self):
-        options = {"particle": "Fickian diffusion", "particle cracking": "negative"}
-        model = pybamm.lithium_ion.DFN(options)
+        options = {"particle cracking": "negative"}
+        model = pybamm.lithium_ion.NewmanTobias(options)
         chemistry = pybamm.parameter_sets.Ai2020
         parameter_values = pybamm.ParameterValues(chemistry=chemistry)
         modeltest = tests.StandardModelTest(model, parameter_values=parameter_values)
         modeltest.test_all()
 
     def test_well_posed_positive_cracking(self):
-        options = {"particle": "Fickian diffusion", "particle cracking": "positive"}
-        model = pybamm.lithium_ion.DFN(options)
+        options = {"particle cracking": "positive"}
+        model = pybamm.lithium_ion.NewmanTobias(options)
         chemistry = pybamm.parameter_sets.Ai2020
         parameter_values = pybamm.ParameterValues(chemistry=chemistry)
         modeltest = tests.StandardModelTest(model, parameter_values=parameter_values)
         modeltest.test_all()
 
     def test_well_posed_both_cracking(self):
-        options = {"particle": "Fickian diffusion", "particle cracking": "both"}
-        model = pybamm.lithium_ion.DFN(options)
+        options = {"particle cracking": "both"}
+        model = pybamm.lithium_ion.NewmanTobias(options)
         chemistry = pybamm.parameter_sets.Ai2020
         parameter_values = pybamm.ParameterValues(chemistry=chemistry)
         modeltest = tests.StandardModelTest(model, parameter_values=parameter_values)